fix: `API` 스펙에 맞게 응답 `dto` 수정
